Move metric tensor and classical jacobian
