+++ pennylane/transforms/__init__.py
@@ -105,8 +105,6 @@ preprocessing, get information from a circuit and more.
 .. autosummary::
     :toctree: api
 
-    ~metric_tensor
-    ~adjoint_metric_tensor
     ~batch_params
     ~batch_input
     ~transforms.insert
@@ -129,7 +127,6 @@ that compute the desired quantity.
 .. autosummary::
     :toctree: api
 
-    ~transforms.classical_jacobian
     ~batch_partial
     ~specs
     ~draw
@@ -192,7 +189,6 @@ from .op_transforms import op_transform
 from .batch_params import batch_params
 from .batch_input import batch_input
 from .batch_partial import batch_partial
-from .classical_jacobian import classical_jacobian
 from .convert_to_numpy_parameters import convert_to_numpy_parameters
 from .compile import compile
 from .decompositions import (
@@ -204,8 +200,6 @@ from .defer_measurements import defer_measurements
 from .sign_expand import sign_expand
 from .hamiltonian_expand import hamiltonian_expand, sum_expand
 from .split_non_commuting import split_non_commuting
-from .metric_tensor import metric_tensor
-from .adjoint_metric_tensor import adjoint_metric_tensor
 from .insert_ops import insert
 
 from .mitigate import mitigate_with_zne, fold_global, poly_extrapolate, richardson_extrapolate
